feat: add genre page, navigation and favorite support (#112)
Ajout de la page Genre avec son ViewModel, services de chargement et édition.
Navigation vers la page d’un genre depuis les pages Artiste, Track et Listening.
Affichage des albums, statistiques et image du genre sur la nouvelle page.
Ajout de la gestion du favori pour les genres (commandes et backend).
Affichage du genre dans les listes de morceaux, avec navigation possible.
Ajout de styles et ressources pour l’affichage du genre.
Corrections de libellés et traductions ("compilations", "Genre").
Refactorisation des types d’identifiants de genre (int → long).
Intégration complète dans la DI et le service de navigation.
Amélioration de l’expérience utilisateur autour des genres musicaux.

+++ Core/Rok.Application/Features/Albums/Query/GetAlbumsByGenreIdQueryHandler.cs
@@ -3,10 +3,10 @@ using Rok.Domain.Interfaces.Entities;
 
 namespace Rok.Application.Features.Albums.Query;
 
-public class GetAlbumsByGenreIdQuery(int genreId) : IQuery<IEnumerable<AlbumDto>>
+public class GetAlbumsByGenreIdQuery(long genreId) : IQuery<IEnumerable<AlbumDto>>
 {
     [RequiredGreaterThanZero]
-    public int GenreId { get; } = genreId;
+    public long GenreId { get; } = genreId;
 }
